% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDeX.R, R/data.R
\docType{data}
\name{cyclegenes}
\alias{cyclegenes}
\title{cyclegenes}
\format{
An object of class \code{data.frame} with 1316 rows and 1 columns.

a data frame with 1316 rows and 1 variables
}
\source{
this data is sourced from the `copycat` package
}
\usage{
cyclegenes
}
\description{
cyclegenes, sourced from copycat package
}
\keyword{datasets}
