% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDeX.R, R/data.R
\docType{data}
\name{full.anno}
\alias{full.anno}
\title{full.anno}
\format{
An object of class \code{data.frame} with 56051 rows and 7 columns.

a data frame with 56051 rows and 7 variables
}
\source{
this data is sourced from the `copycat` package
}
\usage{
full.anno
}
\description{
Annotation file for all known genes, sourced from copycat package
}
\examples{
load(system.file("extdata", "full.anno.RData", package = "SpatialDeX"))
}
\keyword{datasets}
