% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_anno}
\alias{full_anno}
\title{full_anno}
\format{
a data frame with 56051 rows and 7 variables
}
\source{
this data is sourced from the `copycat` package
}
\usage{
full_anno
}
\description{
Annotation file for all known genes, sourced from copycat package
}
\keyword{datasets}
