% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feat_coefficient.R
\name{getsolvedmatrix}
\alias{getsolvedmatrix}
\title{getsolvedmatrix}
\usage{
getsolvedmatrix(Xm, Ym, LAMBDA = 0.01)
}
\arguments{
\item{Xm}{Matrix of predictors.}

\item{Ym}{Matrix of responses.}

\item{LAMBDA}{Regularization parameter,default = 0.01.}
}
\value{
Matrix Amat_g, the solution matrix.
}
\description{
Calculates the solution matrix using ridge regression.
}
\examples{
Xm <- matrix(rnorm(100), nrow=10, ncol=10)
Ym <- matrix(rnorm(10), nrow=10, ncol=1)
getsolvedmatrix(Xm, Ym)
}
