% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feat_coefficient.R
\name{getsolvedmatrix_with_permutation_cell_label}
\alias{getsolvedmatrix_with_permutation_cell_label}
\title{getsolvedmatrix_with_permutation_cell_label}
\usage{
getsolvedmatrix_with_permutation_cell_label(
  Xm,
  Ym,
  LAMBDA = 0.01,
  n_permutation = 10000
)
}
\arguments{
\item{Xm}{Matrix of predictors.}

\item{Ym}{Matrix of responses.}

\item{LAMBDA}{Regularization parameter,default = 0.01.}

\item{n_permutation}{Number of permutations to perform, default = 10,000.}
}
\value{
A list containing two elements:
  \item{Amat_ret}{Matrix of coefficients from the original regression.}
  \item{Amat_ret_higher}{Matrix indicating the proportion of permutations where the absolute value of
  coefficients exceeds those in Amat_ret.}
}
\description{
Computes a solution matrix using ridge regression with the option to permute cell labels.
}
\examples{
Xm <- matrix(rnorm(100), nrow=10, ncol=10)
rownames(Xm) <- paste0("cell_", 1:nrow(Xm))
Ym <- matrix(rnorm(10), nrow=10, ncol=2)
rownames(Ym) <- paste0("cell_", 1:nrow(Ym))
getsolvedmatrix_with_permutation_cell_label(Xm, Ym,n_permutation = 10)
}
