% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDeX.R
\name{replace_smallest_with_zero}
\alias{replace_smallest_with_zero}
\title{replace_smallest_with_zero}
\usage{
replace_smallest_with_zero(row, top_cell_type)
}
\arguments{
\item{row}{a numeric vector}

\item{top_cell_type}{defind the number of cell types to retain based on the
proportion of each cell type. Choose an integer from 0 to 9.}
}
\value{
a numeric vector with the smallest values replaced by zero.
}
\description{
This function replaces the smallest values in a vector with zero.
}
\examples{
row <- c(0.2, 0.1, 0.4, 0.3)
top_cell_type <-2
replace_smallest_with_zero(row,top_cell_type)
}
