% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cyclegenes}
\alias{cyclegenes}
\title{cyclegenes}
\format{
a data frame with 1316 rows and 1 variables
}
\source{
this data is sourced from the `copycat` package
}
\description{
cyclegenes, sourced from copycat package
}
