% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DNA.hg20}
\alias{DNA.hg20}
\title{DNA.hg20}
\format{
a data frame with 12205 rows and 3 variables
}
\source{
this data is sourced from the `copycat` package
}
\usage{
DNA.hg20
}
\description{
DNA.hg20, sourced from copycat package
}
\keyword{datasets}
