% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{full.anno}
\alias{full.anno}
\title{full.anno}
\format{
a data frame with 56051 rows and 7 variables
}
\source{
this data is sourced from the `copycat` package
}
\description{
Annotation file for all known genes, sourced from copycat package
}
