% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDeX.R
\name{SpatialDeX}
\alias{SpatialDeX}
\title{SpatialDeX}
\usage{
SpatialDeX(
  raw_exprs,
  lambda = 1,
  sam.name,
  LOW.DR = 0.03,
  UP.DR = 0.2,
  ngene.chr = 5,
  KS.cut = 0.1,
  win.size = 25,
  top_cell_type = NULL
)
}
\arguments{
\item{raw_exprs}{raw gene expression matrix, with gene symbols in rows and spot names in columns.}

\item{lambda}{lambda}

\item{sam.name}{sample name}

\item{LOW.DR}{minimal population fractions of genes for smoothing for copykat.}

\item{UP.DR}{minimal population fractions of genes for segmentation for copykat.}

\item{ngene.chr}{minimal number of genes per chromosome for cell filtering for copykat.}

\item{KS.cut}{segmentation parameters, input 0 to 1; larger looser criteria for copykat.}

\item{win.size}{minimal window sizes for segmentation for copykat.}

\item{top_cell_type}{defind the number of cell types to retain based on
the proportion of each cell type. Choose an integer from 0 to 9.}
}
\value{
prediction_result_filter
}
\description{
Reference-free cell deconvolution for spatial transcriptome
}
